fix(publish): git rebase strategy option: theirs
from `man git rebase`:
Note that a rebase merge works by replaying each commit from the working branch on top of the <upstream> branch. Because of this, when a merge
conflict happens, the side reported as ours is the so-far rebased series, starting with <upstream>, and theirs is the working branch. In other words,
the sides are swapped.